pm/include/device_clk.h: Fix couple of MISRA.ETYPE.CATEGORY.DIFFERENT.2012
BUILD_ASSERT_OR_ZERO() returns an unsigned but _div is signed and this
causes a:

The operands of the operator '+' do not have the same essential type
category: one is 'signed' and the other is 'unsigned'.

Fixes: SYSFW-1510

+++ pm/include/device_clk.h
@@ -154,7 +154,7 @@ struct dev_clk {
 	  (parent), \
 	.type			= \
 		DEV_CLK_TABLE_TYPE_PARENT, \
-	.div			= (_div) + BUILD_ASSERT_OR_ZERO( \
+	.div			= (_div) + (int) BUILD_ASSERT_OR_ZERO( \
 	  _div < 256), \
 	.idx			= (_idx), \
 	.modify_parent_freq	= \
@@ -165,7 +165,7 @@ struct dev_clk {
 	  (parent), \
 	.type			= \
 		DEV_CLK_TABLE_TYPE_PARENT, \
-	.div			= (_div) + BUILD_ASSERT_OR_ZERO( \
+	.div			= (_div) + (int) BUILD_ASSERT_OR_ZERO( \
 	  _div < 256), \
 	.idx			= (_idx), \
 	.modify_parent_freq	= \
